skype/api/requirements.txt
flask
skpy
pika
python-dotenv
flask_swagger_ui
Werkzeug
pyjwt
gunicorn
requests